extract out the select fn
