+++ web/src/components/devtools/DataCatalog.tsx
@@ -4,9 +4,11 @@ import { RootState } from '../../state/store';
 import {Text, Table, Thead, Tbody, Tfoot, Tr, Th, Td, TableCaption, TableContainer, Badge, Tabs, TabList, Tab, TabPanels, TabPanel} from "@chakra-ui/react";
 
 import { Measure, Dimension } from "web/types";
+import { createAvailableOptions, components } from "../common/SelectComponent";
+import { Select } from 'chakra-react-select';
+import { fetchLayer } from '../../helpers/dataCatalog';
 
 const SLTable = ({table}: {table: Measure[] | Dimension[]}) => {
-  console.log(table, "yoooo", table.length)
   return <TableContainer>
   <Table variant='striped' size="sm" style={{whiteSpace: "pre-wrap"}}>
     <Thead>
@@ -30,10 +32,23 @@ const SLTable = ({table}: {table: Measure[] | Dimension[]}) => {
 export const DataCatalog: React.FC<null> = () => {
   const availableMeasures = useSelector((state: RootState) => state.semanticLayer.availableMeasures) || []
   const availableDimensions = useSelector((state: RootState) => state.semanticLayer.availableDimensions) || []
+  const availableLayers = useSelector((state: RootState) => state.semanticLayer.availableLayers) || []
   const semanticLayer = useSelector((state: RootState) => state.thumbnails.semanticLayer) || ''
 
   return <>
     <Text fontSize="lg" fontWeight="bold">Logic Store: <Badge fontSize={"18px"} variant={"solid"} colorScheme="minusxGreen">{semanticLayer}</Badge></Text>
+    <Select
+      isClearable
+      name={'layers'}
+      options={createAvailableOptions(availableLayers)}
+      placeholder={`Select Semantic Layer`}
+      variant='filled'
+      size={'sm'}
+      value={{ value: semanticLayer, label: semanticLayer }}
+      onChange={fetchLayer}
+      components={components}
+      menuPosition='fixed'
+    />
     <Tabs isFitted variant='enclosed-colored' colorScheme="minusxGreen" mt={5}>
       <TabList mb='1em'>
         <Tab>Measures</Tab>
